requirements.txt
transformers[torch]
tensorflow
biopython
tqdm
tf-keras